+++ src/atomate2/common/analysis/defects/flows.py
@@ -6,18 +6,154 @@ import logging
 from dataclasses import dataclass
 from pathlib import Path
 
-from jobflow import Flow, Maker
+from jobflow import Flow, Job, Maker, OutputReference
 from numpy.typing import NDArray
 from pymatgen.analysis.defects.core import Defect
+from pymatgen.core.structure import Structure
 
 from atomate2.common.analysis.defects.jobs import (
     bulk_supercell_calculation,
+    get_ccd_documents,
+    get_charged_structures,
     get_supercell_from_prv_calc,
     spawn_defect_calcs,
+    spawn_energy_curve_calcs,
 )
 from atomate2.vasp.jobs.core import RelaxMaker
 
 logger = logging.getLogger(__name__)
+
+DEFAULT_DISTORTIONS = (-1, -0.15, -0.1, -0.05, 0, 0.05, 0.1, 0.15, 1)
+
+
+@dataclass
+class ConfigurationCoordinateMaker(Maker):
+    """Maker to generate a configuration coordinate diagram.
+
+    Parameters
+    ----------
+    name: str
+        The name of the flow created by this maker.
+    relax_maker: .BaseVaspMaker or None
+        A maker to perform a atomic-position-only relaxation on the defect charge
+        states.
+    static_maker: .BaseVaspMaker or None
+        A maker to perform the single-shot static calculation of the distorted
+        structures.
+    distortions: tuple[float, ...]
+        The distortions, as a fraction of ΔQ, to use in the calculation of the
+        configuration coordinate diagram.
+    """
+
+    relax_maker: Maker
+    static_maker: Maker
+    name: str = "config. coordinate"
+    # relax_maker: BaseVaspMaker = field(
+    #     default_factory=lambda: RelaxMaker(
+    #         input_set_generator=DEFECT_RELAX_GENERATOR,
+    #     )
+    # )
+    # static_maker: BaseVaspMaker = field(
+    #     default_factory=lambda: StaticMaker(
+    #     input_set_generator=DEFECT_STATIC_GENERATOR)
+    # )
+    distortions: tuple[float, ...] = DEFAULT_DISTORTIONS
+
+    def make(
+        self,
+        structure: Structure,
+        charge_state1: int,
+        charge_state2: int,
+    ):
+        """
+        Make a job for the calculation of the configuration coordinate diagram.
+
+        Parameters
+        ----------
+        structure
+            A structure.
+        charge_state1
+            The reference charge state of the defect.
+        charge_state2
+            The excited charge state of the defect
+
+        Returns
+        -------
+        Flow
+            The full workflow for the calculation of the configuration coordinate
+            diagram.
+        """
+        # use a more descriptive name when possible
+        if not isinstance(structure, OutputReference):
+            name = f"{self.name}: {structure.formula}"
+            if not (
+                isinstance(charge_state1, OutputReference)
+                or isinstance(charge_state2, OutputReference)
+            ):
+                name = (
+                    f"{self.name}: {structure.formula}({charge_state1}-{charge_state2})"
+                )
+
+        # need to wrap this up in a job so that references to undone calculations can
+        # be passed in
+        charged_structures = get_charged_structures(
+            structure, [charge_state1, charge_state2]
+        )
+
+        relax1: Job = self.relax_maker.make(structure=charged_structures.output[0])
+        relax2: Job = self.relax_maker.make(structure=charged_structures.output[1])
+        relax1.append_name(" q1")
+        relax2.append_name(" q2")
+
+        dir1 = relax1.output.dir_name
+        dir2 = relax2.output.dir_name
+        struct1 = relax1.output.structure
+        struct2 = relax2.output.structure
+
+        deformations1 = spawn_energy_curve_calcs(
+            struct1,
+            struct2,
+            distortions=self.distortions,
+            static_maker=self.static_maker,
+            prev_vasp_dir=dir1,
+            add_name="q1",
+            add_info={"relaxed_uuid": relax1.uuid, "distorted_uuid": relax2.uuid},
+        )
+
+        deformations2 = spawn_energy_curve_calcs(
+            struct2,
+            struct1,
+            distortions=self.distortions,
+            static_maker=self.static_maker,
+            prev_vasp_dir=dir2,
+            add_name="q2",
+            add_info={"relaxed_uuid": relax2.uuid, "distorted_uuid": relax1.uuid},
+        )
+
+        deformations1.append_name(" q1")
+        deformations2.append_name(" q2")
+
+        # distortion index with smallest absolute value
+        min_abs_index = min(
+            range(len(self.distortions)), key=lambda i: abs(self.distortions[i])
+        )
+
+        ccd_job = get_ccd_documents(
+            deformations1.output, deformations2.output, undistorted_index=min_abs_index
+        )
+
+        return Flow(
+            jobs=[
+                charged_structures,
+                relax1,
+                relax2,
+                deformations1,
+                deformations2,
+                ccd_job,
+            ],
+            output=ccd_job.output,
+            name=name,
+        )
 
 
 @dataclass
